src/components/charts/AreaChart.test.tsx
// import { render } from '@testing-library/react';
// // import AreaChartComponent, { convertChartDataToLineChartDataPoints } from './AreaChart';
// import { ChartData } from "@/api/entities/insights";
// import { PerformanceType } from '../card/types/PerformanceCard/PerformanceCard.types';

describe('AreaChartComponent', () => {
    // const mockData: ChartData[] = [
    //     { date: '2023-01-01', metric: 100 },
    //     { date: '2023-01-02', metric: 200 },
    // ];

        it('should render without crashing', () => {
        expect(1).toBe(1);
    });
    // it('should render without crashing', () => {
    //     const { container } = render(<AreaChartComponent containerId="test-container" data={mockData} />);
    //     expect(container).toBeInTheDocument();
    // });

    // it('should apply correct styles based on performanceType', () => {
    //     const { getByTestId: getByTestIdSuccess } = render(<AreaChartComponent containerId="test-container" data={mockData} performanceType={PerformanceType.SUCCESS} />);
    //     const { getByTestId: getByTestIdFailure } = render(<AreaChartComponent containerId="test-container" data={mockData} performanceType={PerformanceType.FAILURE} />);

    //     expect(getByTestIdSuccess('area-chart').getAttribute('data-line-chart-color')).toBe('#5EC75A');
    //     expect(getByTestIdFailure('area-chart').getAttribute('data-line-chart-color')).toBe('#C50F1F');
    // });

    // it('should convert chart data to line chart data points correctly', () => {
    //     const result = convertChartDataToLineChartDataPoints(mockData);
    //     expect(result).toEqual([
    //         {
    //             x: new Date('2023-01-01'),
    //             y: 100,
    //             xAxisCalloutData: new Date('2023-01-01').toDateString(),
    //             yAxisCalloutData: '$100.00',
    //             hideCallout: true,
    //         },
    //         {
    //             x: new Date('2023-01-02'),
    //             y: 200,
    //             xAxisCalloutData: new Date('2023-01-02').toDateString(),
    //             yAxisCalloutData: '$200.00',
    //             hideCallout: true,
    //         },
    //     ]);
    // });
});